requirements.txt
discord.py
python-dotenv
aiohttp
astroid
async-timeout
attrs
cffi
chardet
colorama
cryptography
ffmpeg
idna
isort
lazy-object-proxy
mccabe
multidict
pycparser
pylint
pymongo
PyNaCl
pyspnego
six
smbprotocol
toml
typing_extensions
wrapt
yarl
youtube-dl
youtube-search-python

